Rename package `backend` to `offspot_metrics_backend`
