--- backend/src/backend/entrypoint.py
@@ -1,3 +1,0 @@
-from backend.main import create_app  # pragma: no cover
-
-app = create_app()  # pragma: no cover
