+++ backend/src/offspot_metrics_backend/business/agg_kind.py
