+++ backend/src/offspot_metrics_backend/business/indicators/dimensions.py
