+++ backend/src/offspot_metrics_backend/business/indicators/indicator.py
@@ -1,10 +1,10 @@
 import abc
 from typing import Dict, Generator
 
-from backend.business.indicators.dimensions import DimensionsValues
-from backend.business.indicators.holder import Record, State
-from backend.business.indicators.recorder import Recorder
-from backend.business.inputs.input import Input
+from offspot_metrics_backend.business.indicators.dimensions import DimensionsValues
+from offspot_metrics_backend.business.indicators.holder import Record, State
+from offspot_metrics_backend.business.indicators.recorder import Recorder
+from offspot_metrics_backend.business.inputs.input import Input
 
 
 class Indicator(abc.ABC):
