+++ backend/src/offspot_metrics_backend/business/indicators/processor.py
@@ -2,10 +2,10 @@ from typing import List
 
 from sqlalchemy.orm import Session
 
-from backend.business.indicators.indicator import Indicator
-from backend.business.inputs.input import Input
-from backend.business.period import Period
-from backend.db.persister import Persister
+from offspot_metrics_backend.business.indicators.indicator import Indicator
+from offspot_metrics_backend.business.inputs.input import Input
+from offspot_metrics_backend.business.period import Period
+from offspot_metrics_backend.db.persister import Persister
 
 
 class Processor:
