+++ backend/src/offspot_metrics_backend/business/inputs/input.py
