+++ backend/src/offspot_metrics_backend/business/kpis/value.py
