+++ backend/src/offspot_metrics_backend/business/period.py
