+++ backend/src/offspot_metrics_backend/constants.py
