+++ backend/src/offspot_metrics_backend/db/mixins.py
