+++ backend/src/offspot_metrics_backend/entrypoint.py
@@ -1,0 +1,3 @@
+from offspot_metrics_backend.main import create_app  # pragma: no cover
+
+app = create_app()  # pragma: no cover
