+++ backend/src/offspot_metrics_backend/routes/echo.py
