+++ backend/src/simulator.py
@@ -5,11 +5,14 @@ from datetime import datetime, timedelta
 from sqlalchemy import create_engine, delete
 from sqlalchemy.orm import Session
 
-import backend.db.models as dbm
-from backend.business.inputs.content_visit import ContentHomeVisit, ContentObjectVisit
-from backend.business.period import Period
-from backend.business.processor import Processor
-from backend.constants import BackendConf
+import offspot_metrics_backend.db.models as dbm
+from offspot_metrics_backend.business.inputs.content_visit import (
+    ContentHomeVisit,
+    ContentObjectVisit,
+)
+from offspot_metrics_backend.business.period import Period
+from offspot_metrics_backend.business.processor import Processor
+from offspot_metrics_backend.constants import BackendConf
 
 # Some known contents (some are repeated many times to influence the distribution)
 contents = [
