+++ backend/tests/unit/backend/test_constants.py
@@ -1,4 +1,4 @@
-from backend.constants import BackendConf
+from offspot_metrics_backend.constants import BackendConf
 
 
 def test_db_api() -> None:
