+++ backend/tests/unit/backend/test_inheritance.py
@@ -1,9 +1,12 @@
 import pytest
 
-from backend.business.indicators.dimensions import DimensionsValues
-from backend.business.indicators.indicator import Indicator
-from backend.business.indicators.recorder import IntCounterRecorder, Recorder
-from backend.business.inputs.input import Input
+from offspot_metrics_backend.business.indicators.dimensions import DimensionsValues
+from offspot_metrics_backend.business.indicators.indicator import Indicator
+from offspot_metrics_backend.business.indicators.recorder import (
+    IntCounterRecorder,
+    Recorder,
+)
+from offspot_metrics_backend.business.inputs.input import Input
 
 
 class SampleGenericIndicator(Indicator):
