+++ backend/tests/unit/business/indicators/test_indicators.py
@@ -3,14 +3,14 @@ from datetime import datetime, timedelta
 from sqlalchemy import select
 from sqlalchemy.orm import Session
 
-from backend.business.indicators.dimensions import DimensionsValues
-from backend.business.indicators.holder import Record
-from backend.business.indicators.indicator import Indicator
-from backend.business.indicators.processor import Processor
-from backend.business.inputs.input import Input
-from backend.business.period import Period
-from backend.db import count_from_stmt
-from backend.db.models import (
+from offspot_metrics_backend.business.indicators.dimensions import DimensionsValues
+from offspot_metrics_backend.business.indicators.holder import Record
+from offspot_metrics_backend.business.indicators.indicator import Indicator
+from offspot_metrics_backend.business.indicators.processor import Processor
+from offspot_metrics_backend.business.inputs.input import Input
+from offspot_metrics_backend.business.period import Period
+from offspot_metrics_backend.db import count_from_stmt
+from offspot_metrics_backend.db.models import (
     IndicatorDimension,
     IndicatorPeriod,
     IndicatorRecord,
