+++ backend/tests/unit/business/indicators/test_periods.py
@@ -3,12 +3,12 @@ from datetime import datetime
 import pytest
 from sqlalchemy.orm import Session
 
-from backend.business.agg_kind import AggKind
-from backend.business.indicators.indicator import Indicator
-from backend.business.indicators.processor import Processor
-from backend.business.inputs.input import Input
-from backend.business.period import Period
-from backend.db.models import IndicatorPeriod
+from offspot_metrics_backend.business.agg_kind import AggKind
+from offspot_metrics_backend.business.indicators.indicator import Indicator
+from offspot_metrics_backend.business.indicators.processor import Processor
+from offspot_metrics_backend.business.inputs.input import Input
+from offspot_metrics_backend.business.period import Period
+from offspot_metrics_backend.db.models import IndicatorPeriod
 
 
 @pytest.mark.parametrize(
