+++ backend/tests/unit/business/kpis/conftest.py
@@ -4,10 +4,10 @@ from typing import Generator, TypeAlias
 import pytest
 from sqlalchemy.orm import Session
 
-from backend.business.agg_kind import AggKind
-from backend.business.kpis.kpi import Kpi
-from backend.business.kpis.processor import Processor
-from backend.business.period import Period
+from offspot_metrics_backend.business.agg_kind import AggKind
+from offspot_metrics_backend.business.kpis.kpi import Kpi
+from offspot_metrics_backend.business.kpis.processor import Processor
+from offspot_metrics_backend.business.period import Period
 
 ProcessorGenerator: TypeAlias = Generator[Processor, None, None]
 KpiGenerator: TypeAlias = Generator[Kpi, None, None]
